+++ app/src/test/java/com/ar/bootcampar/support/Constants.java
@@ -1,4 +1,4 @@
-package com.ar.bootcampar;
+package com.ar.bootcampar.support;
 
 import com.ar.bootcampar.model.Rol;
