+++ app/src/test/java/com/ar/bootcampar/support/DatabaseSpy.java
@@ -1,4 +1,4 @@
-package com.ar.bootcampar;
+package com.ar.bootcampar.support;
 
 import android.content.ContentValues;
 
@@ -60,3 +60,4 @@ public class DatabaseSpy implements ISQLiteDatabaseWrapper {
 
     public IContentValuesWrapper getInsertedValues() { return insertedValues; }
 }
+
